+++ app/ui/settings/browser.js
@@ -3,12 +3,10 @@ const { ipcRenderer } = require('electron');
 const globalConfig = require('../../lib/global-config.js');
 const clientConfig = require('../../lib/config.js');
 const autoLaunch = require('../../lib/auto-launch.js');
+const tabNavigation = require('../tray/tab-navigation.js');
 
 module.exports = function() {
   function init() {
-    var $navigationItems = $('#settings-navigation li a');
-    var $tabContents = $('.tab-content');
-
     var isLoggedIn = clientConfig.isActiveInstance() !== undefined;
 
     if(process.platform === 'linux') {
@@ -21,20 +19,7 @@ module.exports = function() {
       $('#settings-nav-user').hide();
     }
 
-    navigateTo('settings-global');
-
-    $navigationItems.bind('click', function(event) {
-      event.preventDefault();
-      navigateTo($(this).attr('href').substr(1));
-    });
-
-    function navigateTo(tab) {
-      $navigationItems.parent('li').removeClass('tab-navigation-active');
-      $tabContents.removeClass('tab-active');
-
-      $navigationItems.addBack().find('[href="#' + tab + '"]').parent('li').addClass('tab-navigation-active');
-      $tabContents.addBack().find('#' + tab).addClass('tab-active');
-    }
+    tabNavigation('#settings');
 
     const $autoLaunchCheck = $('#settings-autolaunch-check');
     $autoLaunchCheck.attr('checked', autoLaunch.getState());