+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * CORE: [CHANGE] Add env to instance #24
 * CORE: [CHANGE] Support unlimited quota #109
 * CORE: [FIX] reset instance after server changed #145
+* CORE: [CHANGE] display transfer status in try #103
 
 
 ## 0.2.4